Add files via upload
I finished creating all the IDs in Python and imported the CSV to met db. I have created and inserted data into all the tables and established relationships in SQL. I also connected db to Tableau and put together a dashboard. Feel free to add/remove anything from the dashboard as I am not sure if this is enough or not